Procedural geometry refactoring, createMesh and similar deprecation (#6261)
