+++ examples/src/examples/graphics/reflection-cubemap/example.mjs
@@ -72,7 +72,7 @@ assetListLoader.load(() => {
         app.root.addChild(entity);
 
         // create hight resolution sphere
-        const mesh = pc.createSphere(app.graphicsDevice, { latitudeBands: 200, longitudeBands: 200 });
+        const mesh = pc.Mesh.fromGeometry(app.graphicsDevice, new pc.SphereGeometry({ latitudeBands: 200, longitudeBands: 200 }));
 
         // Add a render component with the mesh
         entity.addComponent('render', {
