+++ src/framework/components/render/component.js
@@ -3,7 +3,7 @@ import { LAYERID_WORLD, RENDERSTYLE_SOLID } from '../../../scene/constants.js';
 import { BatchGroup } from '../../../scene/batching/batch-group.js';
 import { MeshInstance } from '../../../scene/mesh-instance.js';
 import { MorphInstance } from '../../../scene/morph-instance.js';
-import { getShapePrimitive } from '../../../scene/procedural/primitive-cache.js';
+import { getShapePrimitive } from '../../graphics/primitive-cache.js';
 import { GraphNode } from '../../../scene/graph-node.js';
 import { SkinInstanceCache } from '../../../scene/skin-instance-cache.js';
