+++ src/framework/graphics/primitive-cache.js
@@ -1,11 +1,12 @@
 import { Vec2 } from "../../core/math/vec2.js";
-import { createBox } from "./box.js";
-import { createCapsule } from "./capsule.js";
-import { createCone } from "./cone.js";
-import { createPlane } from "./plane.js";
-import { createCylinder } from "./cylinder.js";
-import { createSphere } from "./sphere.js";
-import { createTorus } from "./torus.js";
+import { CapsuleGeometry } from "../../scene/geometry/capsule-geometry.js";
+import { ConeGeometry } from "../../scene/geometry/cone-geometry.js";
+import { CylinderGeometry } from "../../scene/geometry/cylinder-geometry.js";
+import { TorusGeometry } from "../../scene/geometry/torus-geometry.js";
+import { Mesh } from "../../scene/mesh.js";
+import { BoxGeometry } from "../../scene/geometry/box-geometry.js";
+import { SphereGeometry } from "../../scene/geometry/sphere-geometry.js";
+import { PlaneGeometry } from "../../scene/geometry/plane-geometry.js";
 
 // cached mesh primitives
 const shapePrimitives = [];
@@ -28,37 +29,37 @@ function getShapePrimitive(device, type) {
         switch (type) {
 
             case 'box':
-                mesh = createBox(device);
+                mesh = Mesh.fromGeometry(device, new BoxGeometry());
                 area = { x: 2, y: 2, z: 2, uv: (2.0 / 3) };
                 break;
 
             case 'capsule':
-                mesh = createCapsule(device, { radius: 0.5, height: 2 });
+                mesh = Mesh.fromGeometry(device, new CapsuleGeometry({ radius: 0.5, height: 2 }));
                 area = { x: (Math.PI * 2), y: Math.PI, z: (Math.PI * 2), uv: (1.0 / 3 + ((1.0 / 3) / 3) * 2) };
                 break;
 
             case 'cone':
-                mesh = createCone(device, { baseRadius: 0.5, peakRadius: 0, height: 1 });
+                mesh = Mesh.fromGeometry(device, new ConeGeometry({ baseRadius: 0.5, peakRadius: 0, height: 1 }));
                 area = { x: 2.54, y: 2.54, z: 2.54, uv: (1.0 / 3 + (1.0 / 3) / 3) };
                 break;
 
             case 'cylinder':
-                mesh = createCylinder(device, { radius: 0.5, height: 1 });
+                mesh = Mesh.fromGeometry(device, new CylinderGeometry({ radius: 0.5, height: 1 }));
                 area = { x: Math.PI, y: (0.79 * 2), z: Math.PI, uv: (1.0 / 3 + ((1.0 / 3) / 3) * 2) };
                 break;
 
             case 'plane':
-                mesh = createPlane(device, { halfExtents: new Vec2(0.5, 0.5), widthSegments: 1, lengthSegments: 1 });
+                mesh = Mesh.fromGeometry(device, new PlaneGeometry({ halfExtents: new Vec2(0.5, 0.5), widthSegments: 1, lengthSegments: 1 }));
                 area = { x: 0, y: 1, z: 0, uv: 1 };
                 break;
 
             case 'sphere':
-                mesh = createSphere(device, { radius: 0.5 });
+                mesh = Mesh.fromGeometry(device, new SphereGeometry({ radius: 0.5 }));
                 area = { x: Math.PI, y: Math.PI, z: Math.PI, uv: 1 };
                 break;
 
             case 'torus':
-                mesh = createTorus(device, { tubeRadius: 0.2, ringRadius: 0.3 });
+                mesh = Mesh.fromGeometry(device, new TorusGeometry({ tubeRadius: 0.2, ringRadius: 0.3 }));
                 area = { x: Math.PI * 0.5 * 0.5 - Math.PI * 0.1 * 0.1, y: 0.4, z: 0.4, uv: 1 };
                 break;
